ngx-json-log: major rewrite to a single src file
   - version 0.0.8
   - correct request body and response headers handling
   - remove err_ logging
   - encoding operations
   - update documentation
   - travis migration
   - build with unbuntu focal
