--- CHANGES
@@ -1,68 +1,0 @@
-Changes with ngx-json-log 0.0.6                                     27 Aug 2017
-
-    *) hexdump variables
-        *) $http_json_log_req_body_hexdump
-        *) $http_json_err_log_req_hexdump
-    *) $ngx_stream_json_log_payload variable
-    *) log formats moved to main config
-    *) more info while fail to log to file
-    *) fix compilation without stream support
-    *) fix compilation with libjansson < 2.7.0
-    *) arrays zero memory
-    *) kafka bug fixing
-    *) kafka feature check at configure
-    *) code style fixes
-    *) alpine docker and kafka edege update
-
-Changes with ngx-json-log 0.0.5                                     19 Apr 2017
-    *) json log for stream
-    *) json log http bad requests
-       *)  directives log_err_json_*
-    *) kafka
-       *) msg-id from $request_id if available
-       *) optional support if available
-    *) docker
-       *) development docker container -> debian
-       *) runtime docker container -> alpine
-       *) full log format configuration
-       *) elk services
-    *) documentation
-       *) mail log example
-
-Changes with ngx-json-log 0.0.4                                      9 Apr 2017
-    *) stream module log handler
-    *) rename to nxg-json-log
-
-Changes with ngx-http-log-json 0.0.4                                 4 Apr 2017
-    *) module variable $http_log_json_req_headers
-    *) module variable $http_log_json_resp_headers
-    *) module variable $http_log_json_req_body
-    *) local directive http_log_json_req_body_limit
-
-Changes with ngx-http-log-json 0.0.3                                31 Mar 2017
-
-    *) allow to write to multiple destination with differente formats
-    *) http_log_json_format and http_log_json_output directives separation
-    *) significant overhaul for hiding json implementation
-    *) re-factoring for allowing multiple destination/format
-
-Changes with ngx-http-log-json 0.0.2                                26 Mar 2017
-
-    *) upgrade to nginx 1.11.10 support
-    *) parse recipe with pcre regex
-    *) support literal boolean values
-    *) support literal null values
-    *) support array values
-    *) refactoring functions
-    *) unit tests
-    *) travis script integration
-    *) if= condition argument
-
-Changes with ngx-kasha 0.0.1                                        14 Dec 2016
-
-    *) Log output to Kafka topic.
-
-Changes with ngx-kasha 0.0.1                                        11 Dec 2016
-
-    *) The first public version.
-
