--- src/ngx_stream_json_log_module.h
@@ -1,54 +1,0 @@
-/*
- * Copyright (C) 2017 Paulo Pacheco
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS ``AS IS'' AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED.  IN NO EVENT SHALL THE AUTHOR OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS
- * OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
- * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
- * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
- * SUCH DAMAGE.
- */
-#ifndef __NGX_STREAM_JSON_LOG_MODULE_H__
-#define __NGX_STREAM_JSON_LOG_MODULE_H__
-#if nginx_version >= 1011002
-
-#include <ngx_core.h>
-
-#if (NGX_HAVE_LIBRDKAFKA)
-#include "ngx_json_log_kafka.h"
-#endif
-
-struct ngx_stream_json_log_main_conf_s {
-    ngx_array_t                               *formats;
-#if (NGX_HAVE_LIBRDKAFKA)
-    ngx_json_log_main_kafka_conf_t  kafka;
-#endif
-};
-
-typedef struct ngx_stream_json_log_main_conf_s  ngx_stream_json_log_main_conf_t;
-
-struct ngx_stream_json_log_srv_conf_s {
-    ngx_array_t                               *locations;
-};
-
-typedef struct ngx_stream_json_log_srv_conf_s   ngx_stream_json_log_srv_conf_t;
-
-#endif
-
-#endif // __NGX_STREAM_JSON_LOG_MODULE_H__
-
